refactor: move to some functions into util
