+++ tests/integration_test.rs
@@ -14,150 +14,17 @@
  * limitations under the License.
  */
 
+mod util;
+
 use assert_cmd::prelude::*; // Add methods on commands
 use predicates::prelude::*; // Used for writing assertions
-use std::env;
 use std::process::Command; // Run programs
                            // use assert_cmd::cmd::Command; // Run programs - it seems to be equal to "use assert_cmd::prelude::* + use std::process::Command"
 
-use once_cell::sync::Lazy;
-use regex::bytes::Regex;
-use rusoto_core::Region;
-use rusoto_dynamodb::{DynamoDb, DynamoDbClient};
 use std::fs::File;
-use std::io::{self, Write}; // Used when check results by printing to stdout
-use std::sync::Mutex;
-use std::time::Duration;
-use tokio::time::sleep;
+use std::io::Write; // Used when check results by printing to stdout
 
 use tempfile::Builder;
-
-/// Integration tests would go with DynamoDB Local, so before running them setup() starts up DynamoDB Local with Docker.
-/// FYI: https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/DynamoDBLocal.html
-///      https://hub.docker.com/r/amazon/dynamodb-local
-async fn setup() -> Result</* std::process::Command */ Command, Box<dyn std::error::Error>> {
-    setup_with_port(8000).await
-}
-
-// We use std::sync::Mutex instead of tokio::sync::Mutex, because mutex must be poisoned after setup failure.
-static SETUP_MUTEX: Lazy<Mutex<i32>> = Lazy::new(|| Mutex::new(0));
-
-/// Check existence of docker process for dynamodb-local
-fn check_dynamodb_local_running(port: u16) -> bool {
-    let mut docker_for_check = Command::new("docker");
-
-    let check_cmd = docker_for_check.args(&[
-        "ps",
-        "--format",
-        "{{.Ports}}",
-        "--filter",
-        "ancestor=amazon/dynamodb-local",
-    ]);
-    let check_out = check_cmd.output().expect("failed to execute check cmd");
-    let reg_str = format!(r"(?m):{}->\d+/tcp$", port);
-    let port_re = Regex::new(&reg_str).unwrap();
-    if !check_out.status.success() {
-        panic!("failed to execute docker ps command")
-    }
-    port_re.is_match(&check_out.stdout)
-}
-
-async fn setup_with_port(port: i32) -> Result<Command, Box<dyn std::error::Error>> {
-    // Stop docker setup if DYNEIN_TEST_NO_DOCKER_SETUP=true.
-    // This configuration is useful for skipping the docker setup in the GitHub CI environment.
-    // Also, it reduces test time because of skipping of docker checks.
-    // If you use this, you must ensure that docker containers are running for tests.
-    // See https://github.com/awslabs/dynein/pull/59 for detail.
-    let stop_setup: bool = env::var("DYNEIN_TEST_NO_DOCKER_SETUP")
-        .unwrap_or("false".to_string())
-        .to_lowercase()
-        .parse()
-        .expect("DYNEIN_TEST_NO_DOCKER_SETUP expects true or false");
-    if stop_setup {
-        return Ok(Command::cargo_bin("dy")?);
-    }
-
-    // Check the current process at first to allow multiple threads to run tests concurrently.
-    // This is for performance optimization on Windows and Mac OS.
-    // See https://github.com/awslabs/dynein/pull/28#issuecomment-972880324 for detail.
-    if check_dynamodb_local_running(port as u16) {
-        return Ok(Command::cargo_bin("dy")?);
-    };
-
-    // To avoid unnecessary docker container creation, setup docker sequentially
-    let _lock = SETUP_MUTEX.lock();
-
-    // Recheck whether another thread already started the dynamodb-local
-    if check_dynamodb_local_running(port as u16) {
-        return Ok(Command::cargo_bin("dy")?);
-    }
-
-    let mut docker_for_run = Command::new("docker");
-    let docker_run = docker_for_run.args(&[
-        "run",
-        "-p",
-        &format!("{}:8000", port),
-        "-d",
-        "amazon/dynamodb-local",
-    ]);
-    let output = docker_run
-        .output()
-        .expect("failed to running Docker image amazon/dynamodb-local in setup().");
-    if !output.status.success() {
-        panic!("failed to execute docker run command")
-    }
-    print!("DynamoDB Local is up as a container: ");
-    io::stdout().write_all(&output.stdout).unwrap();
-    io::stderr().write_all(&output.stderr).unwrap();
-
-    // Wait dynamodb-local
-    let health_check_url = format!("http://localhost:{}", port);
-    let ddb = DynamoDbClient::new(Region::Custom {
-        name: "local".to_owned(),
-        endpoint: health_check_url,
-    });
-    loop {
-        if let Ok(_result) = ddb.list_tables(Default::default()).await {
-            println!("ListTables API succeeded.");
-            break;
-        } else {
-            println!("Couldn't connect. Retry after 3 seconds.");
-            sleep(Duration::from_secs(3)).await;
-        }
-    }
-
-    Ok(Command::cargo_bin("dy")?)
-}
-
-async fn cleanup(tables: Vec<&str>) -> Result<(), Box<dyn std::error::Error>> {
-    for table in tables {
-        let mut dynein_cmd = setup().await?;
-        let cmd = dynein_cmd.args(&[
-            "--region", "local", "admin", "delete", "table", "--yes", table,
-        ]);
-        cmd.assert().success();
-    }
-    Ok(())
-}
-
-async fn cleanup_with_port(tables: Vec<&str>, port: i32) -> Result<(), Box<dyn std::error::Error>> {
-    for table in tables {
-        let mut dynein_cmd = setup().await?;
-        let cmd = dynein_cmd.args(&[
-            "--region",
-            "local",
-            "--port",
-            &format!("{}", port),
-            "admin",
-            "delete",
-            "table",
-            "--yes",
-            table,
-        ]);
-        cmd.assert().success();
-    }
-    Ok(())
-}
 
 #[tokio::test]
 async fn test_help() -> Result<(), Box<dyn std::error::Error>> {
@@ -169,12 +36,6 @@ async fn test_help() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-async fn cleanup_config(dummy_dir: &str) -> io::Result<()> {
-    use std::fs::remove_dir_all;
-
-    remove_dir_all(dummy_dir)
-}
-
 #[tokio::test]
 async fn test_custom_config_location() -> Result<(), Box<dyn std::error::Error>> {
     use std::path::Path;
@@ -183,7 +44,7 @@ async fn test_custom_config_location() -> Result<(), Box<dyn std::error::Error>>
     let config_dir = dummy_dir.to_string() + "/.dynein";
 
     // cleanup config folder in case it was already there
-    cleanup_config(dummy_dir).await.ok();
+    util::cleanup_config(dummy_dir).await.ok();
 
     // dy config clear
     Command::cargo_bin("dy")?
@@ -195,7 +56,7 @@ async fn test_custom_config_location() -> Result<(), Box<dyn std::error::Error>>
     assert!(Path::new(&config_dir).exists());
 
     // cleanup config folder
-    cleanup_config(dummy_dir).await.ok();
+    util::cleanup_config(dummy_dir).await.ok();
 
     Ok(())
 }
@@ -205,7 +66,7 @@ async fn test_create_table() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_create_table";
 
     // $ dy admin create table <table_name> --keys pk
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let create_cmd = c.args(&[
         "--region", "local", "admin", "create", "table", table_name, "--keys", "pk",
     ]);
@@ -218,7 +79,7 @@ async fn test_create_table() -> Result<(), Box<dyn std::error::Error>> {
         )));
 
     // $ dy admin desc <table_name>
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let desc_cmd = c.args(&["--region", "local", "desc", table_name]);
     desc_cmd
         .assert()
@@ -228,7 +89,7 @@ async fn test_create_table() -> Result<(), Box<dyn std::error::Error>> {
             &table_name
         )));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
@@ -238,7 +99,7 @@ async fn test_create_table_with_region_local_and_port_number_options(
     let table_name = "table--test_create_table_with_region_local_and_port_number_options";
 
     // $ dy admin create table <table_name> --keys pk
-    let mut c = setup_with_port(port).await?;
+    let mut c = util::setup_with_port(port).await?;
     let create_cmd = c.args(&[
         "--region",
         "local",
@@ -260,7 +121,7 @@ async fn test_create_table_with_region_local_and_port_number_options(
         )));
 
     // $ dy admin desc <table_name>
-    let mut c = setup_with_port(port).await?;
+    let mut c = util::setup_with_port(port).await?;
     let desc_cmd = c.args(&[
         "--region",
         "local",
@@ -277,12 +138,12 @@ async fn test_create_table_with_region_local_and_port_number_options(
             &table_name
         )));
 
-    cleanup_with_port(vec![table_name], port).await
+    util::cleanup_with_port(vec![table_name], port).await
 }
 
 #[tokio::test]
 async fn test_scan_non_existent_table() -> Result<(), Box<dyn std::error::Error>> {
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let cmd = c.args(&[
         "--region",
         "local",
@@ -300,56 +161,56 @@ async fn test_scan_non_existent_table() -> Result<(), Box<dyn std::error::Error>
 async fn test_scan_blank_table() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_scan_blank_table";
 
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "admin", "create", "table", table_name, "--keys", "pk",
     ])
     .output()?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let scan_cmd = c.args(&["--region", "local", "--table", table_name, "scan"]);
     scan_cmd
         .assert()
         .success()
         .stdout(predicate::str::contains("No item to show"));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
 async fn test_simple_scan() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_simple_scan";
 
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "admin", "create", "table", table_name, "--keys", "pk",
     ])
     .output()?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&["--region", "local", "--table", table_name, "put", "abc"])
         .output()?;
 
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let scan_cmd = c.args(&["--region", "local", "--table", table_name, "scan"]);
     scan_cmd
         .assert()
         .success()
         .stdout(predicate::str::contains("pk  attributes\nabc"));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 async fn prepare_pk_sk_table(table_name: &&str) -> Result<(), Box<dyn std::error::Error>> {
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "admin", "create", "table", table_name, "--keys", "pk,S", "sk,N",
     ])
     .output()?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "--table", table_name, "put", "abc", "1",
     ])
     .output()?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "--table", table_name, "put", "abc", "2",
     ])
@@ -362,7 +223,7 @@ async fn test_simple_query() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_simple_query";
 
     prepare_pk_sk_table(&table_name).await?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let query_cmd = c.args(&["--region", "local", "--table", table_name, "query", "abc"]);
     query_cmd
         .assert()
@@ -371,7 +232,7 @@ async fn test_simple_query() -> Result<(), Box<dyn std::error::Error>> {
             "pk   sk  attributes\nabc  1\nabc  2",
         ));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
@@ -379,7 +240,7 @@ async fn test_simple_desc_query() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_desc_simple_query";
 
     prepare_pk_sk_table(&table_name).await?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let query_cmd = c.args(&[
         "--region", "local", "--table", table_name, "query", "abc", "-d",
     ]);
@@ -390,7 +251,7 @@ async fn test_simple_desc_query() -> Result<(), Box<dyn std::error::Error>> {
             "pk   sk  attributes\nabc  2\nabc  1",
         ));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
@@ -398,7 +259,7 @@ async fn test_query_limit() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_query_limit";
 
     prepare_pk_sk_table(&table_name).await?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let query_cmd = c.args(&[
         "--region", "local", "--table", table_name, "query", "abc", "-l", "1",
     ]);
@@ -407,14 +268,14 @@ async fn test_query_limit() -> Result<(), Box<dyn std::error::Error>> {
         .success()
         .stdout(predicate::str::contains("pk   sk  attributes\nabc  1"));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
 async fn test_batch_write() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_batch_write";
 
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region", "local", "admin", "create", "table", table_name, "--keys", "pk",
     ])
@@ -461,7 +322,7 @@ async fn test_batch_write() -> Result<(), Box<dyn std::error::Error>> {
         ]
     }
     ")?;
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     c.args(&[
         "--region",
         "local",
@@ -473,7 +334,7 @@ async fn test_batch_write() -> Result<(), Box<dyn std::error::Error>> {
     ])
     .output()?;
 
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let scan_cmd = c.args(&["--region", "local", "--table", table_name, "scan"]);
     scan_cmd
         .assert()
@@ -518,7 +379,7 @@ async fn test_batch_write() -> Result<(), Box<dyn std::error::Error>> {
           "ISBN": "111-1111111111"
         }
     */
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let get_cmd = c.args(&["--region", "local", "--table", table_name, "get", "ichi"]);
     let output = get_cmd.output()?.stdout;
 
@@ -528,7 +389,7 @@ async fn test_batch_write() -> Result<(), Box<dyn std::error::Error>> {
         predicate::str::is_match("\"Dimensions\":")?.eval(String::from_utf8(output)?.as_str())
     );
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
 
 #[tokio::test]
@@ -538,7 +399,7 @@ async fn test_shell_mode() -> Result<(), Box<dyn std::error::Error>> {
     let table_name = "table--test_shell_mode";
 
     // $ dy admin create table <table_name> --keys pk
-    let mut c = setup().await?;
+    let mut c = util::setup().await?;
     let shell_session = c.args(&["--region", "local", "--shell"]);
     let mut tmpfile = Builder::new().tempfile()?.into_file();
     writeln!(tmpfile, "admin create table {} --keys pk", table_name)?;
@@ -554,5 +415,5 @@ async fn test_shell_mode() -> Result<(), Box<dyn std::error::Error>> {
             &table_name
         )));
 
-    cleanup(vec![table_name]).await
+    util::cleanup(vec![table_name]).await
 }
